requirements.txt
streamlit==1.41.1
tensorflow==2.15.1
numpy==1.26.4
pillow==10.4.0
